Implement silk-cni property deny_networks
This allows a platform operator to specify a list of CIDR ranges which
an app container should not be able to talk to, under any circumstances,
including for DNS requests.

This is similar to the garden property (from the garden-runc release)
called "deny_networks", but is applied via silk.

This can be used by a platform operator to limit outbound connectivity
for all apps within an isolation segment, by specifying deny_networks,
including mitigations against DNS exfiltration.

This could, instead, be a feature of application security groups (ASGs)
or egress policies, but is presently infeasible.

Signed-off-by: toby lorne <toby@toby.codes>