fix: silk-daemon-shutdown waits for evacuation
* silk-daemon-shutdown now reads the `containermetadata/store.json` to
check for presence of containers before shutting down. This replaces the
old behavior of checking if the rep server was still up. The end result
is the same, if the rep is finished, there will be no containers on the
cell.
* The check address for the rep server we used to use was turned to
HTTPS. The rep server isn't a public interface that we should be
interacting with. We are removing our dependence on the rep and instead
relying on our own source of truth.
* Discovered as part of [cf-networking-release #76](https://github.com/cloudfoundry/cf-networking-release/issues/76).
This ensures that silk is not prematurely torn down, and applications
stay connected to the overlay network before shutdown.

[#172254086](https://www.pivotaltracker.com/story/show/172254086)

Co-authored-by: Christian Ang <angc@vmware.com>