Delete rules by rule-number when cleaning up parent chains with dynamic-asgs
There are cases where `iptables -S` output cannot be directly translated
back into `iptables -D` commands. If that happens, the Enforce fails,
which has the side-effect of preventing new containers from being
created in garden.

This was run into with rules specifying ICMP packets with type 255, and
code 255. iptables -S uses the "any" shortcut for these rules, but the
rule can only be deleted with the rulespec it was given ("255/255").