Ignore cleanup failures when applying ASG updates
After new ASG chain is created and we fail to clean up old ASG chain we
log the error but continue on and consider our new ASG chain as
successful and save it in the in-memory map for subsequent comparison.
This is to avoid re-applying the same chain. As well as saving it in a
map as a desired chain to avoid deleting it right after we create it.

Original ASG chain fails to be deleted because cascading deletion of
referenced chains might fail. In the observed case we saw that
referenced chain is the netout-...-log chain that is getting referenced
by the new ASG chain now.

Signed-off-by: Marc Paquette <mpaquette@vmware.com>