Ignore cleanup failures when applying ASG updates (#61)
* Ignore cleanup failures when applying ASG updates

After new ASG chain is created and we fail to clean up old ASG chain we
log the error but continue on and consider our new ASG chain as
successful and save it in the in-memory map for subsequent comparison.
This is to avoid re-applying the same chain. As well as saving it in a
map as a desired chain to avoid deleting it right after we create it.

Original ASG chain fails to be deleted because cascading deletion of
referenced chains might fail. In the observed case we saw that
referenced chain is the netout-...-log chain that is getting referenced
by the new ASG chain now.

Signed-off-by: Marc Paquette <mpaquette@vmware.com>

* Keep reject rule in the parent netout chain

When ASG chain is created and jump rule to it is added in the parent
netout chain we clean up the parent netout chain. We want to keep the
REJECT rule in the parent chain so that no traffic is allowed after the
ASG chain in case of any erros while creating ASG chain.

Signed-off-by: Maria Shaldybin <mariash@vmware.com>

* Return the cleanup error when enforcing asg rules

Update local in-memory set with new asg chain if cleanup fails but still
return it as an error. The caller will log it anyway.

Signed-off-by: Marc Paquette <mpaquette@vmware.com>

Co-authored-by: Marc Paquette <mpaquette@vmware.com>