Check last updated endpoint before polling policies (#88)
This PR accompanies cloudfoundry/cf-networking-release#231

Vxlan policy agent first checks the last updated endpoint of policy server and if there was an update (new policy created or deleted) it will run the poll cycle and poll all policies for containers on Diego cell. It will also run poll cycle if getting policies returns an error for backwards compatibility with old policy servers.

It only does this in the converger and not when forcing poll cycle during container creation.