Use container guid in ASG chain name
Using shortened sha1 is prone to collisions.

Instead of using sha1 of container guids, getting 6 characters and
adding timestamp use container guid up to 24 characters. Similar to how
netout and overlay chains are created. The new format:

asg-{container-guid[:20]}

In this case container-guid is stripped from dashes and `check` prefix
used in health check containers.

To replace ASG chain, a temporary chain is created with prefix `casg-`
(candidate ASG chain). It is populated with new rules and appened to
parent chain (netout). Then original ASG chain is deleted. And `casg-`
chain is renamed to `asg-` chain. This is done to avoid incorrect state
if any of operations fail.

force-orphaned-asgs-cleanup endpoint now will be deleting ASG chain by
container guid looking specifically for chain name and candidate chain
name instead of using regex. Since we are not using timestamps and regex
is not needed.

Signed-off-by: Andy Nguyen <andy-a.nguyen@broadcom.com>
Signed-off-by: Krsna Mahapatra <krsna.mahapatra@broadcom.com>
Signed-off-by: Maria Shaldybin <maria.shaldybin@broadcom.com>