Replacing asg chain does not try to delete jump targets
Jump targets created in ASG chain are netout--handle--log
chains when iptables logging is turned on. When we replace the chain
with another chain, these targets might be used still. Instead of
getting a bunch of device is busy errors and keep retrying the update,
skip deleting jump targets if they are in use by candidate chain.