Update silk-controller for multi-cidr overlay network
Now the silk-controller will give out leases from all of the CIDRs
provided in the overlay network. The first block of the first CIDR is
reserved for SingleIP usage. The first block of the other CIDRs are also
reserved to setup the networking between CIDRs.

For example, the bosh properties...
* network: ["10.255.0.0/23", "10.255.10.0/23", "10.255.20.0/22"]
* subnet_prefix_length: 24

Will result in...
* 10.255.0.0/24 --> reserved for single IP + networking setup
* 10.255.1.0/24 --> possible lease for a Diego Cell

* 10.255.10.0/24 --> reserved for networking setup
* 10.255.11.0/24 --> possible lease for a Diego Cell

* 10.255.20.0/24 --> reserved for networking setup
* 10.255.21.0/24 --> possible lease for a Diego Cell
* 10.255.22.0/24 --> possible lease for a Diego Cell
* 10.255.23.0/24 --> possible lease for a Diego Cell