silk-controller job templates work for multi-cidr overlay networks
Now the silk-controller network property can be an array of CIDRs.
Previously it was a single string of one CIDR.  ow the template creates
a config with an array for this property, whether or not the operator
provided an array or a single string.

Previously allowed example:
* "10.255.0.0/16"

Now allowed examples:
* "10.255.0.0/16" (still allowed for backwards compatibility)
* ["10.255.0.0/16"]
* ["10.50.0.0/23", "10.100.0.0/22", "10.250.0.0/23"]